requirements.txt
streamlit==1.31.1
typer==0.9.0
apscheduler==3.10.4
loguru==0.7.2

sqlalchemy==2.0.27
sqlite-utils==3.35

uvicorn==0.27.0.post1
fastapi==0.109.2

langchain-core>=1.0
langchain-google-genai>=3.1.0
google-genai
python-dotenv
llama-index
